nnue rewrite bench 3659421
